+++ roseau/load_flow/exceptions.py
@@ -41,14 +41,15 @@ class RoseauLoadFlowExceptionCode(StrEnum):
     BAD_Y_SHUNT_VALUE = auto()
     BAD_TYPE_NAME_SYNTAX = auto()
     BAD_LENGTH_VALUE = auto()
-    BAD_MAX_CURRENTS_VALUE = auto()
-    BAD_MAX_CURRENTS_SIZE = auto()
+    BAD_AMPACITIES_VALUE = auto()
+    BAD_AMPACITIES_SIZE = auto()
     BAD_SECTIONS_VALUE = auto()
     BAD_SECTIONS_SIZE = auto()
     BAD_MATERIALS_VALUE = auto()
     BAD_MATERIALS_SIZE = auto()
     BAD_INSULATORS_VALUE = auto()
     BAD_INSULATORS_SIZE = auto()
+    BAD_MAX_LOADING_VALUE = auto()
 
     # Transformer
     BAD_TRANSFORMER_ID = auto()